+++ lib/src/flutter_luban.dart
@@ -1,24 +1,25 @@
 import 'dart:async';
-import 'dart:io';
 import 'dart:isolate';
 
+import 'package:cross_file/cross_file.dart';
 import 'package:flutter/foundation.dart';
+import 'package:flutter_luban/src/types/compress_object.dart';
 import 'package:flutter_luban/src/utils/calculator.dart';
 import 'package:flutter_luban/src/utils/compressor.dart';
 import 'package:flutter_luban/src/utils/image_parser.dart';
+import 'package:flutter_luban/src/utils/xfile/interface.dart';
+import 'package:flutter_luban/src/utils/xfile/io_file_storage.dart';
+import 'package:flutter_luban/src/utils/xfile/web_file_storage.dart';
 import 'package:image/image.dart';
-
-import 'types/compress_object.dart';
 
 class Luban {
   static final CompressionCalculator _calculator = CompressionCalculator();
+  static final XFileStorage _xFileStorage =
+      kIsWeb ? WebFileStorage() : IoXFileStorage();
 
   Luban._();
 
-  static Future<String?> compressImage(CompressObject object) async {
-    if (kIsWeb) {
-      throw "Because the web does not support isolate, it does not support web image compression for the time being";
-    }
+  static Future<XFile?> compressImage(CompressObject object) async {
     return compute(_lubanCompress, object);
   }
 
@@ -31,7 +32,7 @@ class Luban {
     return answer.first;
   }
 
-  static Future<List<String?>> compressImageList(
+  static Future<List<XFile?>> compressImageList(
       List<CompressObject> objects) async {
     return compute(_lubanCompressList, objects);
   }
@@ -46,34 +47,53 @@ class Luban {
     });
   }
 
-  static List<String?> _lubanCompressList(List<CompressObject> objects) {
-    var results = [];
-    objects.forEach((_o) {
-      results.add(_lubanCompress(_o));
+  static Future<List<XFile?>> _lubanCompressList(
+      List<CompressObject> objects) async {
+    var results = <XFile?>[];
+    await Future.forEach(objects, (obj) async {
+      results.add(await _lubanCompress(obj));
     });
-    return results as List<String?>;
+    return results;
   }
 
-  static Future<String?> _lubanCompress(CompressObject object) async {
-    Image? image = decodeImage(object.imageFile.readAsBytesSync());
+  static Future<XFile?> _lubanCompress(CompressObject object) async {
+    final originalFileName = object.imageXFile.path.split("/").last;
+    String targetPath = object.targetPath ?? _xFileStorage.tempDir();
+    final imageFormat =
+        ImageParser.parseFormat(object.toRgb, object.imageXFile.path);
+    XFile? decodedImageFile;
+    String path =
+        '$targetPath/luban_$originalFileName.${ImageParser.getSuffix(imageFormat)}';
+    decodedImageFile = XFile(path);
+
+    if (await _xFileStorage.exists(decodedImageFile)) {
+      if (object.useCache) {
+        return decodedImageFile;
+      }
+      _xFileStorage.delete(decodedImageFile);
+    }
+    final bytes = await _lubanCompressToBytes(object.imageXFile,
+        toRgb: object.toRgb, numberOfColors: object.numberOfColors);
+    if (bytes == null) return null;
+    decodedImageFile = XFile.fromData(bytes);
+    if (!kIsWeb) await decodedImageFile.saveTo(path);
+    return decodedImageFile;
+  }
+
+  static Future<Uint8List?> _lubanCompressToBytes(XFile imageXFile,
+      {bool toRgb = false, int numberOfColors = 128}) async {
+    final bool exists = await _xFileStorage.exists(imageXFile);
+    if (!exists) {
+      return null;
+    }
+    final imageBytes = await imageXFile.readAsBytes();
+    Image? image = decodeImage(imageBytes);
     if (image == null) {
       throw Exception("flutter_luban don't support this image type");
     }
-    ImageFormat format = ImageFormat.jpg;
-    if (object.toJpg) {
-      image = ImageParser.convertToJpg(image) ?? image;
-    } else {
-      format = ImageParser.parseFormat(object.imageFile.path);
-    }
-    final originalFileName = object.imageFile.path.split("/").last;
-    File? decodedImageFile;
-    decodedImageFile = File(
-        '${object.targetPath}/luban_$originalFileName.${ImageParser.getSuffix(format)}');
-    if (decodedImageFile.existsSync()) {
-      if (object.useCache) {
-        return decodedImageFile.path;
-      }
-      decodedImageFile.deleteSync();
+    final format = ImageParser.parseFormat(toRgb, imageXFile.path);
+    if (format == ImageFormat.jpg) {
+      image = ImageParser.convertToRgb(image);
     }
     final target = _calculator.calculateTarget(image.width, image.height);
     if (target.width > 0 && target.width < image.width) {
@@ -83,12 +103,12 @@ class Luban {
         height: target.height,
       );
     }
-    final data = Compressor.compress(
+    final bytes = Compressor.compress(
       image,
       targetSizeKb: target.targetSizeKb,
       imageFormat: format,
+      numberOfColors: numberOfColors,
     );
-    decodedImageFile.writeAsBytesSync(data);
-    return decodedImageFile.path;
+    return bytes;
   }
 }